sprint03: Minor fix t11

+++ sprint03/t11/mx_is_mersenne.c
@@ -4,7 +4,7 @@ bool mx_is_prime(int num);
 double mx_pow(double n, unsigned int pow);
 
 bool mx_is_mersenne(int n) {
-    int pow = 0;
+    double pow = 0;
 
     if (!mx_is_prime(n)) {
         return false;
